+++ nextgisweb/vector_layer/model.py
@@ -1,6 +1,8 @@
 # -*- coding: utf-8 -*-
 from __future__ import division, absolute_import, print_function, unicode_literals
+
 import json
+from logging import error
 import uuid
 import zipfile
 import ctypes
@@ -8,7 +10,7 @@ from datetime import datetime, time, date
 import six
 
 from zope.interface import implementer
-from osgeo import ogr, osr
+from osgeo import gdal, ogr, osr
 from shapely.geometry import box
 from sqlalchemy.sql import ColumnElement
 from sqlalchemy.ext.compiler import compiles
@@ -23,6 +25,7 @@ from sqlalchemy import (
 
 from ..event import SafetyEvent
 from .. import db
+from ..core.exception import ValidationError
 from ..resource import (
     Resource,
     DataScope,
@@ -31,12 +34,12 @@ from ..resource import (
     SerializedProperty as SP,
     SerializedRelationship as SR,
     ResourceGroup)
-from ..resource.exception import ValidationError, ResourceError
+from ..resource.exception import ResourceError
 from ..env import env
 from ..models import declarative_base, DBSession, migrate_operation
 from ..layer import SpatialLayerMixin, IBboxLayer
 from ..lib.geometry import Geometry
-from ..compat import lru_cache
+from ..compat import lru_cache, html_escape
 
 from ..feature_layer import (
     Feature,
@@ -96,8 +99,53 @@ _FIELD_TYPE_2_ENUM = dict(zip(FIELD_TYPE_OGR, FIELD_TYPE.enum))
 _FIELD_TYPE_2_DB = dict(zip(FIELD_TYPE.enum, FIELD_TYPE_DB))
 
 SCHEMA = 'vector_layer'
+DRIVERS_SUPPORTED = ('ESRI Shapefile', 'GeoJSON', 'KML', 'GML')
+OPEN_OPTIONS = ('EXPOSE_FID=NO', )
 
 Base = declarative_base(dependencies=('resource', 'feature_layer'))
+
+
+def translate(trstring):
+    return env.core.localizer().translate(trstring)
+
+
+class ERROR_FIX(object):
+    NONE = 'NONE'
+    SAFE = 'SAFE'
+    LOSSY = 'LOSSY'
+
+    default = NONE
+    enum = (NONE, SAFE, LOSSY)
+
+
+skip_errors_default = False
+
+error_limit = 10
+
+
+class TOGGLE(object):
+    AUTO = None
+    YES = True
+    NO = False
+
+    enum = (AUTO, YES, NO)
+
+
+geom_cast_params_default = dict(
+    geometry_type=TOGGLE.AUTO,
+    is_multi=TOGGLE.AUTO,
+    has_z=TOGGLE.AUTO)
+
+
+class FID_SOURCE(object):
+    AUTO = 'AUTO'
+    SEQUENCE = 'SEQUENCE'
+    FIELD = 'FIELD'
+
+
+fid_params_default = dict(
+    fid_source=FID_SOURCE.SEQUENCE,
+    fid_field=None)
 
 
 class FieldDef(object):
@@ -122,50 +170,122 @@ class TableInfo(object):
         self.metadata = None
         self.table = None
         self.model = None
+        self.fid_field_index = None
         self.fmap = None
+        self.geometry_type = None
 
     @classmethod
-    def from_ogrlayer(cls, ogrlayer, srs_id, strdecode):
+    def from_ogrlayer(cls, ogrlayer, srs_id, strdecode, skip_other_geometry_types,
+                      fid_params, geom_cast_params):
         self = cls(srs_id)
 
+        if geom_cast_params['geometry_type'] == GEOM_TYPE.POINT:
+            geom_filter = set(GEOM_TYPE.points)
+        elif geom_cast_params['geometry_type'] == GEOM_TYPE.LINESTRING:
+            geom_filter = set(GEOM_TYPE.linestrings)
+        elif geom_cast_params['geometry_type'] == GEOM_TYPE.POLYGON:
+            geom_filter = set(GEOM_TYPE.polygons)
+        else:
+            geom_filter = set(GEOM_TYPE.enum)
+
+        if geom_cast_params['is_multi'] == TOGGLE.NO:
+            geom_filter -= set(GEOM_TYPE.is_multi)
+        elif geom_cast_params['is_multi'] == TOGGLE.YES:
+            geom_filter = geom_filter.intersection(set(GEOM_TYPE.is_multi))
+
+        if geom_cast_params['has_z'] == TOGGLE.NO:
+            geom_filter -= set(GEOM_TYPE.has_z)
+        elif geom_cast_params['has_z'] == TOGGLE.YES:
+            geom_filter = geom_filter.intersection(set(GEOM_TYPE.has_z))
+
         ltype = ogrlayer.GetGeomType()
-        self.geometry_type = _GEOM_OGR_2_TYPE[ltype]
 
-        is_multi = self.geometry_type in GEOM_TYPE.is_multi
-        has_z = self.geometry_type in GEOM_TYPE.has_z
+        if len(geom_filter) == 1:
+            self.geometry_type = geom_filter.pop()
+        elif ltype in GEOM_TYPE_OGR and _GEOM_OGR_2_TYPE[ltype] in geom_filter:
+            self.geometry_type = _GEOM_OGR_2_TYPE[ltype]
+        elif len(geom_filter) > 1:
+            is_multi = False
+            has_z = False
 
-        if not is_multi or not has_z:
             for feature in ogrlayer:
-                geom = feature.GetGeometryRef()
-                gtype = geom.GetGeometryType()
-                if ltype != gtype:
-                    if not is_multi and (
-                        ltype in (ogr.wkbPoint, ogr.wkbPoint25D) and gtype in (ogr.wkbMultiPoint, ogr.wkbMultiPoint25D)  # NOQA: E501
-                        or ltype in (ogr.wkbLineString, ogr.wkbLineString25D) and gtype in (ogr.wkbMultiLineString, ogr.wkbMultiLineString25D)  # NOQA: E501, W503
-                        or ltype in (ogr.wkbPolygon, ogr.wkbPolygon25D) and gtype in (ogr.wkbMultiPolygon, ogr.wkbMultiPolygon25D)  # NOQA: E501, W503
-                    ):
-                        self.geometry_type = _GEOM_OGR_2_TYPE[gtype]
-                        ltype = gtype
-                        is_multi = True
-
-                    if not has_z and (
-                        ltype in (ogr.wkbPoint, ogr.wkbMultiPoint) and gtype in (ogr.wkbPoint25D, ogr.wkbMultiPoint25D)  # NOQA: E501
-                        or ltype in (ogr.wkbLineString, ogr.wkbMultiLineString) and gtype in (ogr.wkbLineString25D, ogr.wkbMultiLineString25D)  # NOQA: E501, W503
-                        or ltype in (ogr.wkbPolygon, ogr.wkbMultiPolygon) and gtype in (ogr.wkbPolygon25D, ogr.wkbMultiPolygon25D)  # NOQA: E501, W503
-                    ):
-                        self.geometry_type = _GEOM_OGR_2_TYPE[gtype]
-                        ltype = gtype
-                        has_z = True
-
-                if is_multi and has_z:
+                if len(geom_filter) <= 1:
                     break
 
+                geom = feature.GetGeometryRef()
+                if geom is None:
+                    continue
+                gtype = geom.GetGeometryType()
+                if gtype in (ogr.wkbGeometryCollection, ogr.wkbGeometryCollection25D) \
+                    and geom.GetGeometryCount() == 1:
+                    geom = geom.GetGeometryRef(0)
+                    gtype = geom.GetGeometryType()
+
+                if gtype not in GEOM_TYPE_OGR:
+                    continue
+                geometry_type = _GEOM_OGR_2_TYPE[gtype]
+
+                if geom_cast_params['geometry_type'] == TOGGLE.AUTO:
+                    if geometry_type in GEOM_TYPE.points:
+                        geom_filter = geom_filter.intersection(set(GEOM_TYPE.points))
+                    elif geometry_type in GEOM_TYPE.linestrings:
+                        geom_filter = geom_filter.intersection(set(GEOM_TYPE.linestrings))
+                    elif geometry_type in GEOM_TYPE.polygons:
+                        geom_filter = geom_filter.intersection(set(GEOM_TYPE.polygons))
+                elif skip_other_geometry_types and geometry_type not in geom_filter:
+                    continue
+
+                if geom_cast_params['is_multi'] == TOGGLE.AUTO and not is_multi \
+                    and geometry_type in GEOM_TYPE.is_multi:
+                    geom_filter = geom_filter.intersection(set(GEOM_TYPE.is_multi))
+                    is_multi = True
+
+                if geom_cast_params['has_z'] == TOGGLE.AUTO and not has_z \
+                    and geometry_type in GEOM_TYPE.has_z:
+                    geom_filter = geom_filter.intersection(set(GEOM_TYPE.has_z))
+                    has_z = True
+
+            if geom_cast_params['is_multi'] == TOGGLE.AUTO and not is_multi:
+                geom_filter = geom_filter - set(GEOM_TYPE.is_multi)
+
+            if geom_cast_params['has_z'] == TOGGLE.AUTO and not has_z:
+                geom_filter = geom_filter - set(GEOM_TYPE.has_z)
+
             ogrlayer.ResetReading()
+
+            if len(geom_filter) == 1:
+                self.geometry_type = geom_filter.pop()
+
+        if self.geometry_type is None:
+            err_msg = _("Could not determine a geometry type.")
+            if len(geom_filter) == 0:
+                err_msg += " " + _("Source layer contains no features satisfying the conditions.")
+            raise VE(message=err_msg)
 
         self.fields = []
 
         defn = ogrlayer.GetLayerDefn()
+
+        if fid_params['fid_source'] in (FID_SOURCE.AUTO, FID_SOURCE.FIELD):
+            if fid_params['fid_field'] is not None:
+                idx = defn.GetFieldIndex(fid_params['fid_field'])
+                if idx != -1:
+                    fld_defn = defn.GetFieldDefn(idx)
+                    if fld_defn.GetType() == ogr.OFTInteger:
+                        self.fid_field_index = idx
+
+            if self.fid_field_index is None and fid_params['fid_source'] == FID_SOURCE.FIELD:
+                if fid_params['fid_field'] is None:
+                    raise VE(_("Parameter 'fid_field' is missing."))
+                else:
+                    if idx == -1:
+                        raise VE(_("Field '%s' not found.") % fid_params['fid_field'])
+                    else:
+                        raise VE(_("Field '%s' type is not integer.") % fid_params['fid_field'])
+
         for i in range(defn.GetFieldCount()):
+            if i == self.fid_field_index:
+                continue
             fld_defn = defn.GetFieldDefn(i)
 
             # TODO: Fix invalid field names as done for attributes.
@@ -283,9 +403,14 @@ class TableInfo(object):
                 for k, v in six.iteritems(kwargs):
                     setattr(self, k, v)
 
+        sequence = db.Sequence(tablename + '_id_seq', start=1,
+                               minvalue=-2**31, metadata=metadata)
         table = db.Table(
             tablename,
-            metadata, db.Column('id', db.Integer, primary_key=True),
+            metadata, db.Column(
+                'id', db.Integer,
+                sequence,
+                primary_key=True),
             db.Column('geom', ga.Geometry(
                 dimension=2, srid=self.srs_id,
                 geometry_type=geom_fldtype)),
@@ -296,53 +421,171 @@ class TableInfo(object):
         db.mapper(model, table)
 
         self.metadata = metadata
+        self.sequence = sequence
         self.table = table
         self.model = model
         self.fmap = {fld.keyname: fld.key for fld in self.fields}
 
-    def load_from_ogr(self, ogrlayer, strdecode):
+    def load_from_ogr(self, ogrlayer, strdecode, skip_other_geometry_types,
+                      fix_errors, skip_errors):
         source_osr = ogrlayer.GetSpatialRef()
         target_osr = osr.SpatialReference()
         target_osr.ImportFromEPSG(self.srs_id)
 
         transform = osr.CoordinateTransformation(source_osr, target_osr)
 
-        is_multi = self.geometry_type in GEOM_TYPE.is_multi
-        has_z = self.geometry_type in GEOM_TYPE.has_z
+        errors = []
 
-        for fid, feature in enumerate(ogrlayer):
+        max_fid = None
+        for i, feature in enumerate(ogrlayer, start=1):
+            if len(errors) >= error_limit:
+                break
+
+            if self.fid_field_index is None:
+                fid = i
+            else:
+                fid = feature.GetFieldAsInteger(self.fid_field_index)
+            max_fid = max(max_fid, fid) if max_fid is not None else fid
+
             geom = feature.GetGeometryRef()
+            if geom is None:
+                if not skip_other_geometry_types:
+                    errors.append(_("Feature #%d doesn't have geometry.") % fid)
+                continue
+
+            # Extract GeometryCollection
+            if geom.GetGeometryType() in (ogr.wkbGeometryCollection, ogr.wkbGeometryCollection25D) \
+               and fix_errors != ERROR_FIX.NONE:
+                geom_candidate = None
+                for i in range(geom.GetGeometryCount()):
+                    col_geom = geom.GetGeometryRef(i)
+                    col_gtype = col_geom.GetGeometryType()
+                    if col_gtype not in GEOM_TYPE_OGR:
+                        continue
+                    if (self.geometry_type in GEOM_TYPE.points and col_gtype in (
+                           ogr.wkbPoint, ogr.wkbPoint25D,
+                           ogr.wkbMultiPoint, ogr.wkbMultiPoint25D)) \
+                       or (self.geometry_type in GEOM_TYPE.linestrings and col_gtype in (
+                           ogr.wkbLineString, ogr.wkbLineString25D,
+                           ogr.wkbMultiLineString, ogr.wkbMultiLineString25D)) \
+                       or (self.geometry_type in GEOM_TYPE.polygons and col_gtype in (
+                           ogr.wkbPolygon, ogr.wkbPolygon25D,
+                           ogr.wkbMultiPolygon, ogr.wkbMultiPolygon25D)):
+                        if geom_candidate is None:
+                            geom_candidate = col_geom
+                            if fix_errors == ERROR_FIX.LOSSY:
+                                break
+                        else:
+                            errors.append(_("Feature #%d has multiple geometries satisfying the conditions.") % fid)
+                            continue
+                if geom_candidate is not None:
+                    geom = geom_candidate
 
             gtype = geom.GetGeometryType()
-            if gtype not in GEOM_TYPE_OGR:
-                raise ValidationError(_(
-                    "Unknown geometry type: %d (%s).") % (
-                    gtype, ogr.GeometryTypeToName(gtype)))
 
-            geom.Transform(transform)
+            # Check geometry type
+            if gtype not in GEOM_TYPE_OGR:
+                if not skip_other_geometry_types:
+                    errors.append(_(
+                        "Feature #%d has unknown geometry type: %d (%s).") % (
+                        fid, gtype, ogr.GeometryTypeToName(gtype)))
+                continue
+            elif not any((
+                (self.geometry_type in GEOM_TYPE.points
+                    and _GEOM_OGR_2_TYPE[gtype] in GEOM_TYPE.points),
+                (self.geometry_type in GEOM_TYPE.linestrings
+                    and _GEOM_OGR_2_TYPE[gtype] in GEOM_TYPE.linestrings),
+                (self.geometry_type in GEOM_TYPE.polygons
+                    and _GEOM_OGR_2_TYPE[gtype] in GEOM_TYPE.polygons),
+            )):
+                if not skip_other_geometry_types:
+                    errors.append(_(
+                        "Feature #%d has unsuitable geometry type: %d (%s).") % (
+                        fid, gtype, ogr.GeometryTypeToName(gtype)))
+                continue
 
             # Force single geometries to multi
-            if is_multi:
+            if self.geometry_type in GEOM_TYPE.is_multi:
                 if gtype in (ogr.wkbPoint, ogr.wkbPoint25D):
                     geom = ogr.ForceToMultiPoint(geom)
                 elif gtype in (ogr.wkbLineString, ogr.wkbLineString25D):
                     geom = ogr.ForceToMultiLineString(geom)
                 elif gtype in (ogr.wkbPolygon, ogr.wkbPolygon25D):
                     geom = ogr.ForceToMultiPolygon(geom)
+            elif gtype in (ogr.wkbMultiPoint, ogr.wkbMultiPoint25D,
+                           ogr.wkbMultiLineString, ogr.wkbMultiLineString25D,
+                           ogr.wkbMultiPolygon, ogr.wkbMultiPolygon25D):
+                if geom.GetGeometryCount() == 1 or fix_errors == ERROR_FIX.LOSSY:
+                    geom = geom.GetGeometryRef(0)
+                else:
+                    errors.append(_("Feature #%d has multiple geometries satisfying the conditions.") % fid)
+                    continue
+
+            geom.Transform(transform)
 
             # Force Z
+            has_z = self.geometry_type in GEOM_TYPE.has_z
             if has_z and not geom.Is3D():
                 geom.Set3D(True)
+            elif not has_z and geom.Is3D():
+                geom.Set3D(False)
+          
+            # Points can't have validity errors.
+            is_single = _GEOM_OGR_2_TYPE[gtype] not in GEOM_TYPE.is_multi
+            is_point = _GEOM_OGR_2_TYPE[gtype] in GEOM_TYPE.points
+            is_polygon = _GEOM_OGR_2_TYPE[gtype] in GEOM_TYPE.polygons
 
-            # Check geometry type again
-            gtype = geom.GetGeometryType()
-            if gtype not in GEOM_TYPE_OGR or _GEOM_OGR_2_TYPE[gtype] != self.geometry_type:
-                raise ValidationError(_(
-                    "Unknown geometry type: %d (%s).") % (
-                    gtype, ogr.GeometryTypeToName(gtype)))
+            if not is_point and not geom.IsValid():
+                # Close rings for polygons: GDAL doesn't provide a method for
+                # checking if a geometry has unclosed rings, but we can achieve
+                # this via comparison.
+                if is_polygon:
+                    geom_closed = geom.Clone()
+                    geom_closed.CloseRings()
+                    if not geom_closed.Equals(geom):
+                        if fix_errors == ERROR_FIX.NONE:
+                            errors.append(_("Feature #%d has unclosed rings.") % fid)
+                            continue
+                        else:
+                            geom = geom_closed
+
+                # Check for polygon rings with fewer than 3 points and
+                # linestrings with fewer than 2 points.
+                if not geom.IsValid():
+                    error_found = False
+                    for part in ((geom, ) if is_single else geom):
+                        if is_polygon:
+                            for ring in part:
+                                if ring.GetPointCount() < 4:
+                                    # TODO: Invalid parts can be removed from multipart geometries in LOSSY mode.
+                                    errors.append(_("Feature #%d has less than 3 points in a polygon ring.") % fid)
+                                    error_found = True
+                        elif part.GetPointCount() < 2:
+                            # TODO: Invalid parts can be removed from multipart geometries in LOSSY mode.
+                            errors.append(_("Feature #%d has less than 2 points in a linestring.") % fid)
+                            error_found = True
+                        if error_found:
+                            break
+                    if error_found:
+                        continue
+                
+                # NOTE: Disabled for better times.
+                # Check for topology errors and fix them as possible.
+                # invalid = True
+                # if fix_errors != ERROR_FIX.NONE:
+                #     if fix_errors == ERROR_FIX.LOSSY and not geom.IsValid():
+                #         geom = geom.MakeValid()
+                #         if geom is not None and not geom.IsValid():
+                #             geom = geom.Buffer(0)
+                #     invalid = geom is None or not geom.IsValid() or geom.GetGeometryType() != gtype
+                # if invalid:
+                #     errors.append(_("Feature #%d has invalid geometry.") % fid)
+                #     continue
 
             fld_values = dict()
             for i in range(feature.GetFieldCount()):
+                if i == self.fid_field_index:
+                    continue
                 fld_type = feature.GetFieldDefnRef(i).GetType()
 
                 if (not feature.IsFieldSet(i) or feature.IsFieldNull(i)):
@@ -377,20 +620,34 @@ class TableInfo(object):
                     try:
                         fld_value = strdecode(feature.GetFieldAsString(i))
                     except UnicodeDecodeError:
-                        raise ValidationError(_(
+                        errors.append(_(
                             "It seems like declared and actual attributes "
                             "encodings do not match. Unable to decode "
                             "attribute #%(attr)d of feature #%(feat)d. "
                             "Try declaring different encoding.") % dict(
                             feat=fid, attr=i))
+                        continue
 
                 fld_name = strdecode(feature.GetFieldDefnRef(i).GetNameRef())
                 fld_values[self[fld_name].key] = fld_value
 
-            obj = self.model(fid=fid, geom=ga.elements.WKBElement(
+            if len(errors) > 0 and not skip_errors:
+                continue
+
+            obj = self.model(id=fid, geom=ga.elements.WKBElement(
                 bytearray(geom.ExportToWkb(ogr.wkbNDR)), srid=self.srs_id), **fld_values)
 
             DBSession.add(obj)
+
+        if len(errors) > 0 and not skip_errors:
+            detail = '<br>'.join(html_escape(translate(error)) for error in errors)
+            raise VE(message=_("Vector layer cannot be written due to errors."), detail=detail)
+
+        # Set sequence next value
+        if max_fid is not None:
+            connection = DBSession.connection()
+            connection.execute('ALTER SEQUENCE "%s"."%s" RESTART WITH %d' %
+                               (self.sequence.schema, self.sequence.name, max_fid + 1))
 
 
 class VectorLayerField(Base, LayerField):
@@ -436,8 +693,13 @@ class VectorLayer(Base, Resource, SpatialLayerMixin, LayerFieldsMixin):
     def _tablename(self):
         return 'layer_%s' % self.tbl_uuid
 
-    def setup_from_ogr(self, ogrlayer, strdecode):
-        tableinfo = TableInfo.from_ogrlayer(ogrlayer, self.srs.id, strdecode)
+    def setup_from_ogr(self, ogrlayer, strdecode=lambda x: x,
+                       skip_other_geometry_types=False,
+                       fid_params=fid_params_default,
+                       geom_cast_params=geom_cast_params_default):
+        tableinfo = TableInfo.from_ogrlayer(
+            ogrlayer, self.srs.id, strdecode, skip_other_geometry_types,
+            fid_params, geom_cast_params)
         tableinfo.setup_layer(self)
 
         tableinfo.setup_metadata(self._tablename)
@@ -455,8 +717,11 @@ class VectorLayer(Base, Resource, SpatialLayerMixin, LayerFieldsMixin):
 
         self.tableinfo = tableinfo
 
-    def load_from_ogr(self, ogrlayer, strdecode):
-        self.tableinfo.load_from_ogr(ogrlayer, strdecode)
+    def load_from_ogr(self, ogrlayer, strdecode=lambda x: x,
+                      skip_other_geometry_types=False,
+                      fix_errors=ERROR_FIX.default, skip_errors=skip_errors_default):
+        self.tableinfo.load_from_ogr(
+            ogrlayer, strdecode, skip_other_geometry_types, fix_errors, skip_errors)
 
     def get_info(self):
         return super(VectorLayer, self).get_info() + (
@@ -744,7 +1009,29 @@ VE = ValidationError
 
 class _source_attr(SP):
 
-    def _ogrds(self, ogrds):
+    def _ogrds(self, filename, encoding):
+        if six.PY2:
+            with _set_encoding(encoding) as sdecode:
+                ogrds = gdal.OpenEx(filename, 0, allowed_drivers=DRIVERS_SUPPORTED, open_options=OPEN_OPTIONS)
+                strdecode = sdecode
+        else:
+            # Ignore encoding option in Python 3
+            ogrds = gdal.OpenEx(filename, 0, allowed_drivers=DRIVERS_SUPPORTED, open_options=OPEN_OPTIONS)
+
+            def strdecode(x):
+                return x
+
+        if ogrds is None:
+            ogrds = ogr.Open(filename, 0)
+            if ogrds is None:
+                raise VE(_("GDAL library failed to open file."))
+            else:
+                drivername = ogrds.GetDriver().GetName()
+                raise VE(_("Unsupport OGR driver: %s.") % drivername)
+
+        return ogrds, strdecode
+
+    def _ogrlayer(self, ogrds):
         if ogrds.GetLayerCount() < 1:
             raise VE(_("Dataset doesn't contain layers."))
 
@@ -757,59 +1044,68 @@ class _source_attr(SP):
 
         return ogrlayer
 
-    def _ogrlayer(self, obj, ogrlayer, recode=lambda a: a):
+    def _setup_layer(self, obj, ogrlayer, skip_other_geometry_types, fix_errors, skip_errors,
+                     geom_cast_params, fid_params, strdecode):
         if ogrlayer.GetSpatialRef() is None:
             raise VE(_("Layer doesn't contain coordinate system information."))
-
-        for feat in ogrlayer:
-            geom = feat.GetGeometryRef()
-            if geom is None:
-                raise VE(_("Feature #%d doesn't have geometry.") % feat.GetFID())
-
-        ogrlayer.ResetReading()
 
         obj.tbl_uuid = uuid.uuid4().hex
 
         with DBSession.no_autoflush:
-            obj.setup_from_ogr(ogrlayer, recode)
-            obj.load_from_ogr(ogrlayer, recode)
+            obj.setup_from_ogr(ogrlayer, strdecode, skip_other_geometry_types, fid_params, geom_cast_params)
+            obj.load_from_ogr(ogrlayer, strdecode, skip_other_geometry_types, fix_errors, skip_errors)
 
     def setter(self, srlzr, value):
         if srlzr.obj.id is not None:
             raise ValidationError("Source parameter does not apply to update vector layer.")
 
         datafile, metafile = env.file_upload.get_filename(value['id'])
-        encoding = value.get('encoding', 'utf-8')
+
+        encoding = srlzr.data.get('encoding')
+        if encoding is None:
+            # backward compatibility
+            encoding = value.get('encoding', 'utf-8')
 
         iszip = zipfile.is_zipfile(datafile)
         ogrfn = ('/vsizip/{%s}' % datafile) if iszip else datafile
 
-        if six.PY2:
-            with _set_encoding(encoding) as sdecode:
-                ogrds = ogr.Open(ogrfn, 0)
-                recode = sdecode
-        else:
-            # Ignore encoding option in Python 3
-            ogrds = ogr.Open(ogrfn, 0)
+        ogrds, strdecode = self._ogrds(ogrfn, encoding)
+        ogrlayer = self._ogrlayer(ogrds)
 
-            def recode(x):
-                return x
+        skip_other_geometry_types = srlzr.data.get('skip_other_geometry_types')
 
-        if ogrds is None:
-            raise VE(_("GDAL library failed to open file."))
+        fix_errors = srlzr.data.get('fix_errors', ERROR_FIX.default)
+        if fix_errors not in ERROR_FIX.enum:
+            raise VE(_("Unknown 'fix_errors' value."))
 
-        drivername = ogrds.GetDriver().GetName()
+        skip_errors = srlzr.data.get('skip_errors', skip_errors_default)
 
-        if drivername not in ('ESRI Shapefile', 'GeoJSON', 'KML', 'GML'):
-            raise VE(_("Unsupport OGR driver: %s.") % drivername)
+        geometry_type = srlzr.data.get(
+            'cast_geometry_type', geom_cast_params_default['geometry_type'])
+        if geometry_type not in (None, 'POINT', 'LINESTRING', 'POLYGON'):
+            raise VE(_("Unknown 'cast_geometry_type' value."))
 
-        ogrlayer = self._ogrds(ogrds)
-        geomtype = ogrlayer.GetGeomType()
-        if geomtype not in _GEOM_OGR_2_TYPE:
-            raise VE(_("Unsupported geometry type: '%s'. Probable reason: data contain mixed geometries.") % (  # NOQA: E501
-                ogr.GeometryTypeToName(geomtype) if geomtype is not None else None))
+        is_multi = srlzr.data.get('cast_is_multi', geom_cast_params_default['is_multi'])
+        if is_multi not in TOGGLE.enum:
+            raise VE(_("Unknown 'cast_is_multi' value."))
 
-        self._ogrlayer(srlzr.obj, ogrlayer, recode)
+        has_z = srlzr.data.get('cast_has_z', geom_cast_params_default['has_z'])
+        if has_z not in TOGGLE.enum:
+            raise VE(_("Unknown 'cast_has_z' value."))
+
+        geom_cast_params = dict(
+            geometry_type=geometry_type,
+            is_multi=is_multi,
+            has_z=has_z)
+
+        fid_source = srlzr.data.get('fid_source', fid_params_default['fid_source'])
+        fid_params = dict(
+            fid_source=fid_source,
+            fid_field=srlzr.data.get('fid_field')
+        )
+
+        self._setup_layer(srlzr.obj, ogrlayer, skip_other_geometry_types, fix_errors, skip_errors,
+                          geom_cast_params, fid_params, strdecode)
 
 
 class _fields_attr(SP):
@@ -845,9 +1141,10 @@ class VectorLayerSerializer(Serializer):
     resclass = VectorLayer
 
     srs = SR(read=P_DSS_READ, write=P_DSS_WRITE)
-    geometry_type = _geometry_type_attr(read=P_DSS_READ, write=P_DSS_WRITE)
 
     source = _source_attr(read=None, write=P_DS_WRITE)
+
+    geometry_type = _geometry_type_attr(read=P_DSS_READ, write=P_DSS_WRITE)
     fields = _fields_attr(read=None, write=P_DS_WRITE)